refactoring: Free the submodules under ./lib
Putting all the package submodules under ./lib seems pretty weird
according to someone with much larger golang expertise than me.

I'm blindly following his nitpicks as I want to trust his criteria
rather than mine.
