+++ cmd/subcmds/delete.go
@@ -2,8 +2,8 @@ package subcmds
 
 import (
 	"fmt"
-	"github.com/bgartzi/uhhmm/lib/inventory"
-	"github.com/bgartzi/uhhmm/lib/local"
+	"github.com/bgartzi/uhhmm/inventory"
+	"github.com/bgartzi/uhhmm/local"
 	"github.com/urfave/cli/v2"
 	"strconv"
 )
