+++ inventory/inventory.go
@@ -1,7 +1,7 @@
 package inventory
 
 import (
-	"github.com/bgartzi/uhhmm/lib/host"
+	"github.com/bgartzi/uhhmm/host"
 )
 
 type InventoryBackend interface {
