+++ local/keys.go
@@ -6,9 +6,9 @@ import (
 	"crypto/x509"
 	"encoding/pem"
 	"fmt"
-	"github.com/bgartzi/uhhmm/lib/host"
-	"github.com/bgartzi/uhhmm/lib/uhhm"
-	"github.com/bgartzi/uhhmm/lib/utils"
+	"github.com/bgartzi/uhhmm/host"
+	"github.com/bgartzi/uhhmm/uhhm"
+	"github.com/bgartzi/uhhmm/utils"
 	"golang.org/x/crypto/ssh"
 	"io/ioutil"
 	"os"
